Move CloudURL to httpClient.ts
Integration tests returned this error:

  TypeError: Class extends value undefined is not a constructor or null

A quick search revealed that this is often caused by circular
dependencies, which turned out to be true.

Madge[0] turned out to be a great help to find those.

This is what it looked like:
1. All sub-clients (i.e. `Notifier`) import `HTTPClient`
2. HTTPClient imports `CloudURL` from `Client`
3. Client imports all sub-clients

This was fixed by moving the `CloudURL` const to `HTTPClient`.

[0]: https://github.com/pahen/madge